feat: AI가이드 추천 루트 생성 API 구현

+++ src/main/java/com/mey/backend/domain/route/entity/Route.java
@@ -48,7 +48,7 @@ public class Route extends BaseTimeEntity {
     private double totalDistance;
 
     @Column(nullable = false)
-    private double totalCost;
+    private int totalCost;
 
     @JdbcTypeCode(SqlTypes.JSON)
     @Column(nullable = false, columnDefinition = "json")
